pyproject.toml
[tool.poetry]
name = "calibrate-google-trends"
version = "0.1.0"
description = "This project fork from GoogleTrendsAnchorBank with the aim to integrate with mongodb."
authors = ["Hsiu-Hsuan Yeh"]
license = "MIT"
readme = "README.md"
packages = [{include = "gtab"}]


[tool.poetry.dependencies]
python = "^3.11"
hydra-core = "^1.3.2"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
tqdm = "^4.66.1"
networkx = "^3.2.1"
pytrends = "^4.9.2"
numpy = "^1.26.4"
pandas = "^2.2.0"
pymongo = "^4.6.1"
googlesearch-python = "^1.2.3"
httpx = "^0.26.0"
keybert = "^0.8.3"
orjson = "^3.9.13"
selectolax = "^0.3.19"
urllib3 = "==1.*"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.0"
ipywidgets = "^8.1.1"
jupyter-bokeh = "^3.0.7"
hvplot = "^0.9.1"
pytest = "^7.4.4"
jupyterlab-execute-time = "2.3.1"
jupyterlab = "3.6.6"
jupyterlab-lsp = "4.3.0"
python-lsp-server = {extras = ["all"], version = "^1.8.2"}
poetry-kernel = "^0.1.3"
sphinx = "^7.2.6"
sphinx-rtd-theme = "^2.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
